requirements.txt
#https://github.com/martinblech/xmltodict
xmltodict
# web server with Flask and Bootstrap4
pyFlaskBootstrap4>=0.5.1
flask-dropzone
# https://github.com/SmartDataAnalytics/OpenResearch
ConferenceCorpus>=0.0.26
# https://pypi.org/project/pylodstorage/
pylodstorage>=0.4.6
# http://pydispatcher.sourceforge.net/
PyDispatcher
# https://pypi.org/project/APScheduler/
APScheduler
# https://github.com/pyparsing/pyparsing
pyparsing>=3.0.9